lint: fix accidental double-promotion warning

+++ src/addons/stats.c
@@ -62,7 +62,7 @@ void print_value(
     float value)
 {
     ecs_size_t len = ecs_os_strlen(name);
-    printf("%s: %*s %.2f\n", name, 32 - len, "", value);
+    printf("%s: %*s %.2f\n", name, 32 - len, "", (double)value);
 }
 
 static
@@ -128,7 +128,7 @@ void ecs_get_world_stats(
     record_counter(&s->pipeline_build_count_total, t, world->stats.pipeline_build_count_total);
     record_counter(&s->systems_ran_frame, t, world->stats.systems_ran_frame);
 
-    if (delta_world_time != 0.0 && delta_frame_count != 0.0) {
+    if (delta_world_time != 0.0f && delta_frame_count != 0.0f) {
         record_gauge(
             &s->fps, t, 1.0f / (delta_world_time / (float)delta_frame_count));
     } else {
